Task 1: the project runs
The project now has it's dependencies managed by maven, and can run on a Tomcat container (the container used originally) without editing the project in any way. To run simply make sure that the tomcat server has https activated (this site does not work with http) and upload the mia-applicazione.war file that is in the target table to the container. (the project did not have all of is functionality accessible and that was not fixed so some pages must be accessed by changing the url in the browser bar directly)